fix(macos): build fixes from the first build on a Mac
- PanelPlacement (and its test) import CoreGraphics where available: on
  macOS CGRect's height/midX/initialisers are not visible through
  Foundation alone.
- Double <-> CGFloat conversions at the HeadroomKit boundary are explicit.
  CGFloat is a distinct type on macOS and the implicit conversion does not
  apply to optionals (AccountList drop indicator) or frame parameters.
- bundle.sh stages the helper as dist/headroom-daemon: on a
  case-insensitive file system dist/headroom overwrote the app binary
  dist/Headroom. Inside the bundle it stays Contents/Helpers/headroom.

+++ shell/macos/Sources/HeadroomUI/MenuBarLabel.swift
@@ -35,7 +35,7 @@
             ZStack {
                 Circle().stroke(Color.black.opacity(0.25), lineWidth: 2.5)
                 Circle()
-                    .trim(from: 0, to: min(1, max(0, fraction)))
+                    .trim(from: 0, to: CGFloat(min(1, max(0, fraction))))
                     .stroke(Color.black, style: StrokeStyle(lineWidth: 2.5, lineCap: .round))
                     .rotationEffect(.degrees(-90))
             }
